+++ src/locale/EditTexts.ts
@@ -14,13 +14,13 @@ type EditTexts = {
     /** [formatted] A way to say "empty line between [node1] and [node2]" */
     line: Template<['before', 'after']>;
     /** [formatted] A description of how many conflicts are at this position */
-    conflicts: Template<['count']>;
+    conflicts: Template<['#count']>;
     /** [plain] Says the selected delimiter's matching partner is on the same line. $name: the partner delimiter's label */
     delimiterMatchedSameLine: Template<['name']>;
     /** [plain] Says the selected delimiter's matching partner is some lines below. $name: partner label, $lines: number of lines */
-    delimiterMatchedBelow: Template<['name', 'lines']>;
+    delimiterMatchedBelow: Template<['name', '#lines']>;
     /** [plain] Says the selected delimiter's matching partner is some lines above. $name: partner label, $lines: number of lines */
-    delimiterMatchedAbove: Template<['name', 'lines']>;
+    delimiterMatchedAbove: Template<['name', '#lines']>;
     /** [plain] Says the selected delimiter has no matching partner */
     delimiterUnmatched: string;
     /** [formatted] $1: node description */